alarm database implemented

+++ app/src/main/java/com/example/rosadowning/nocrastinate/BroadcastReceivers/MidnightDataResetReceiver.java
@@ -9,6 +9,7 @@ import android.content.SharedPreferences;
 import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
 
+import com.example.rosadowning.nocrastinate.DBHelpers.AlarmDBContract;
 import com.example.rosadowning.nocrastinate.DBHelpers.StatsDBContract;
 import com.example.rosadowning.nocrastinate.DBHelpers.ToDoReaderContract;
 import com.example.rosadowning.nocrastinate.DataModels.StatsIconData;
@@ -27,61 +28,69 @@ public class MidnightDataResetReceiver extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
 
-        Log.d(TAG, "Midnight Receiver reached");
+        Log.d(TAG, "Midnight Receiver Reached");
 
         ReentrantLock reentrantLock = new ReentrantLock();
         StatsIconData midnightStat = new StatsIconData();
 
-        SharedPreferences bootPreferences = context.getSharedPreferences("BootData", Context.MODE_PRIVATE);
-        SharedPreferences.Editor bootEditor = bootPreferences.edit();
-
-        try{
-            reentrantLock.lock();
-            bootEditor.putLong("MidnightReceiver", System.currentTimeMillis());
-        } finally {
-            reentrantLock.unlock();
-        }
-
         DateTime today = new DateTime().withTimeAtStartOfDay();
         Date yesterday = today.minusDays(1).withTimeAtStartOfDay().toDate();
 
+        boolean alarmAlreadySet = false;
         long overallTime;
         long tasksCompleted;
         int unlocks;
 
         try {
             reentrantLock.lock();
+            AlarmDBContract.AlarmDBHelper alarmDBHelper = new AlarmDBContract.AlarmDBHelper(context);
+            SQLiteDatabase sqlRead = alarmDBHelper.getReadableDatabase();
 
-            ToDoReaderContract.ToDoListDbHelper toDoHelper = new ToDoReaderContract.ToDoListDbHelper(context);
-            SQLiteDatabase sqlToDo = toDoHelper.getReadableDatabase();
-            tasksCompleted = toDoHelper.getNoOfCompletedToDos(yesterday.getTime(), today.getMillis());
-
-            SharedPreferences sharedPreferences = context.getSharedPreferences("StatisticsInfo", Context.MODE_PRIVATE);
-            SharedPreferences.Editor editor = sharedPreferences.edit();
-            overallTime = sharedPreferences.getLong("totalDuration", 0);
-            unlocks = sharedPreferences.getInt("noOfUnlocks", 0);
-
-            editor.clear();
-            editor.commit();
-
+            if (alarmDBHelper.isAlarmSet(today.getMillis())) {
+                alarmAlreadySet = true;
+            } else {
+                SQLiteDatabase sqlWrite = alarmDBHelper.getWritableDatabase();
+                alarmDBHelper.insertAlarm(today.getMillis());
+            }
         } finally {
             reentrantLock.unlock();
         }
 
-        midnightStat.setDate(yesterday);
-        midnightStat.setOverallTime(overallTime);
-        midnightStat.setNoOfUnlocks(unlocks);
-        midnightStat.setTasksCompleted(tasksCompleted);
+        if (!alarmAlreadySet) {
 
-        StatsDBContract.StatsDBHelper statsHelper = new StatsDBContract.StatsDBHelper(context);
-        SQLiteDatabase sqlStats = statsHelper.getWritableDatabase();
-        statsHelper.insertNewStat(midnightStat);
+            try {
+                reentrantLock.lock();
 
-        DateTime tomorrow = today.plusDays(1).withTimeAtStartOfDay();
-        Intent midnightIntent = new Intent(context, MidnightDataResetReceiver.class);
-        PendingIntent startPIntent = PendingIntent.getBroadcast(context, 0, midnightIntent, 0);
-        AlarmManager alarm = (AlarmManager) context.getSystemService(ALARM_SERVICE);
-        alarm.setExactAndAllowWhileIdle(AlarmManager.RTC_WAKEUP, tomorrow.getMillis(), startPIntent);
+                ToDoReaderContract.ToDoListDbHelper toDoHelper = new ToDoReaderContract.ToDoListDbHelper(context);
+                SQLiteDatabase sqlToDo = toDoHelper.getReadableDatabase();
+                tasksCompleted = toDoHelper.getNoOfCompletedToDos(yesterday.getTime(), today.getMillis());
 
+                SharedPreferences sharedPreferences = context.getSharedPreferences("StatisticsInfo", Context.MODE_PRIVATE);
+                SharedPreferences.Editor editor = sharedPreferences.edit();
+                overallTime = sharedPreferences.getLong("totalDuration", 0);
+                unlocks = sharedPreferences.getInt("noOfUnlocks", 0);
+
+                editor.clear();
+                editor.commit();
+
+            } finally {
+                reentrantLock.unlock();
+            }
+
+            midnightStat.setDate(yesterday);
+            midnightStat.setOverallTime(overallTime);
+            midnightStat.setNoOfUnlocks(unlocks);
+            midnightStat.setTasksCompleted(tasksCompleted);
+
+            StatsDBContract.StatsDBHelper statsHelper = new StatsDBContract.StatsDBHelper(context);
+            SQLiteDatabase sqlStats = statsHelper.getWritableDatabase();
+            statsHelper.insertNewStat(midnightStat);
+
+            DateTime tomorrow = today.withTimeAtStartOfDay().plusDays(1);
+            Intent midnightIntent = new Intent(context, MidnightDataResetReceiver.class);
+            PendingIntent startPIntent = PendingIntent.getBroadcast(context, 0, midnightIntent, 0);
+            AlarmManager alarm = (AlarmManager) context.getSystemService(ALARM_SERVICE);
+            alarm.setExactAndAllowWhileIdle(AlarmManager.RTC_WAKEUP, tomorrow.getMillis(), startPIntent);
+        }
     }
 }
